requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
python-whois==0.8.0
dnspython==2.4.2
pyOpenSSL==23.3.0
beautifulsoup4==4.12.2
